py_backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.2.0
httpx>=0.27.0
gTTS>=2.5.1
aiofiles>=23.2.1
fast-alpr>=0.4.0
